backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
pandas==2.1.3
openpyxl==3.1.2

# ✅ NUEVAS: Dependencias de seguridad
slowapi==0.1.9                    # Rate limiting
python-decouple==3.8              # Environment variables seguras 